% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newspaper-package.R
\docType{package}
\name{newspaper-package}
\alias{newspaper}
\alias{newspaper-package}
\title{newspaper: Newspaper's Contents Extracter}
\description{
Contents Extracter from Newspaper's website.
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
