% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_newspaper.R
\name{newspaper_install}
\alias{newspaper_install}
\title{Set up the Python environment for the newspaper package}
\usage{
newspaper_install()
}
\value{
Invisibly returns \code{NULL}
}
\description{
This function installs \link[reticulate:install_miniconda]{miniconda} if it is
not present, then sets up a \link[reticulate:conda-tools]{conda environment}
with the tools needed to run \code{\link[=newspaper_scrape]{newspaper_scrape()}}.
}
\examples{
\dontrun{newspaper_install()}
}
